My SQL Practice using Jupyter notebook
This notebook contains SQL code which is to analyze the overall activities of space missions over a period. 
I wrote the SQL codes on jupyter notebook so as to enable me understand and also explain my ideas
Finally, the sql flavour i used here is mysql which is just my personal preference.